main: return early if cmd endpoint response is None
With the Home Assistant WebSocket command endpoint, the response is sent
back to Willow in a callback instead of in the original request. We do
this by having the command endpoint return None in the parse_response
function. Now that we added inline WAC in the command endpoint handling
code, we should make sure we don't run WAC code when not needed.